Added Snake AI; Torch 0.2!
Snakes now run away from lights and "attack" players. However, they also
attack torches and each other because AI functions do not have access to
the World object, and thus cannot lookup the type of the entities around
them. This should be fixed when the World object is finally dissolved.